Split Parakeet ownership interleaving tests
